+++ src/main/java/cech12/unlitcampfire/mixinaccess/ICampfireBlockMixin.java
@@ -6,8 +6,13 @@ import net.minecraft.world.level.block.state.properties.BooleanProperty;
 public interface ICampfireBlockMixin {
 
     BooleanProperty INFINITE = BooleanProperty.create("infinite");
+    BooleanProperty RUNS_OUT = BooleanProperty.create("runs_out");
 
     int getMaxLitTime(BlockState state);
+
+    int unlitCampfire$getMaxLitTimeExtension(BlockState state);
+
+    int unlitCampfire$getRunsOutIndicatorTime(BlockState state);
 
     boolean burnsInfinite(BlockState state);