+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Forge Recommended Versioning](https://mcforge.readthedocs.io/en/latest/conventions/versioning/).
 
+## [1.20.1-1.9.1.0] - 2024-09-14
+### Added
+- new behaviour that a (soul) campfire decreases the light level for the last 30 seconds (configurable, can be deactivated) (all loaders) (thanks to Acenthus for the idea) #38
+- new configuration options "campfireMaxLitTimeExtension" and "soulCampfireMaxLitTimeExtension" to extend the burn time beyond the normal burn time when adding burnables (all loaders) (thanks to Fer0x453 for the idea) #42
+
 ## [1.20.1-1.9.0.3] - 2024-07-18
 ### Fixed
 - (soul) campfires refused to go out by rain when they were configured to burn infinitely (burn time = 0) (all loaders) (thanks to Enchilada for the report) #40